+++ src/components/popover-container/popover-container.component.tsx
@@ -3,6 +3,7 @@ import { PaddingProps } from "styled-system";
 import { Transition, TransitionStatus } from "react-transition-group";
 import { flip, offset } from "@floating-ui/dom";
 
+import useMediaQuery from "../../hooks/useMediaQuery";
 import {
   PopoverContainerWrapperStyle,
   PopoverContainerHeaderStyle,
@@ -17,6 +18,8 @@ import createGuid from "../../__internal__/utils/helpers/guid";
 import { filterStyledSystemPaddingProps } from "../../style/utils";
 import useClickAwayListener from "../../hooks/__internal__/useClickAwayListener";
 import Events from "../../__internal__/utils/helpers/events";
+import FocusTrap from "../../__internal__/focus-trap";
+import { ModalContext } from "../modal";
 
 export interface RenderOpenProps {
   tabIndex: number;
@@ -123,6 +126,8 @@ export interface PopoverContainerProps extends PaddingProps {
   closeButtonAriaLabel?: string;
   /** Container aria label */
   containerAriaLabel?: string;
+  /** Disables the animation for the component */
+  disableAnimation?: boolean;
 }
 
 const popoverMiddleware = [
@@ -148,6 +153,7 @@ export const PopoverContainer = ({
   openButtonAriaLabel,
   closeButtonAriaLabel = "close",
   containerAriaLabel,
+  disableAnimation = false,
   ...rest
 }: PopoverContainerProps) => {
   const isControlled = open !== undefined;
@@ -164,16 +170,17 @@ export const PopoverContainer = ({
 
   const isOpen = isControlled ? open : isOpenInternal;
 
-  useEffect(() => {
-    if (isOpen && closeButtonRef.current)
-      setTimeout(() => closeButtonRef.current?.focus(), 0);
-  }, [isOpen]);
+  const reduceMotion = !useMediaQuery(
+    "screen and (prefers-reduced-motion: no-preference)"
+  );
 
   const closePopover = useCallback(
     (ev: React.MouseEvent<HTMLElement> | React.KeyboardEvent<HTMLElement>) => {
-      if (!isControlled) setIsOpenInternal(!isOpen);
+      if (!isControlled) setIsOpenInternal(false);
       if (onClose) onClose(ev);
-      if (isOpen && openButtonRef.current) openButtonRef.current.focus();
+      if (isOpen && openButtonRef.current) {
+        openButtonRef.current.focus();
+      }
     },
     [isControlled, isOpen, onClose]
   );
@@ -188,7 +195,9 @@ export const PopoverContainer = ({
         );
       });
 
-      if (!eventIsFromSelectInput && Events.isEscKey(ev)) closePopover(ev);
+      if (!eventIsFromSelectInput && Events.isEscKey(ev)) {
+        closePopover(ev);
+      }
     },
     [closePopover]
   );
@@ -219,7 +228,7 @@ export const PopoverContainer = ({
   };
 
   const renderOpenComponentProps = {
-    tabIndex: isOpen ? -1 : 0,
+    tabIndex: 0,
     "aria-expanded": isOpen,
     "aria-haspopup": "dialog" as const,
     isOpen,
@@ -244,6 +253,67 @@ export const PopoverContainer = ({
   };
 
   const handleClick = useClickAwayListener(handleClickAway, "mousedown");
+  const [isAnimationComplete, setIsAnimationComplete] = useState(false);
+
+  const popover = (
+    <Transition
+      in={isOpen}
+      timeout={{ exit: 300 }}
+      appear
+      mountOnEnter
+      unmountOnExit
+      nodeRef={popoverContentNodeRef}
+      onEntered={
+        shouldCoverButton
+          ? /* istanbul ignore next */ () => setIsAnimationComplete(true)
+          : undefined
+      }
+      onExiting={
+        shouldCoverButton
+          ? /* istanbul ignore next */ () => setIsAnimationComplete(false)
+          : undefined
+      }
+    >
+      {(state: TransitionStatus) =>
+        isOpen && (
+          <Popover
+            disablePortal
+            reference={popoverReference}
+            placement={position === "right" ? "bottom-start" : "bottom-end"}
+            popoverStrategy={
+              disableAnimation || reduceMotion ? "fixed" : "absolute"
+            }
+            {...(shouldCoverButton && { middleware: popoverMiddleware })}
+          >
+            <PopoverContainerContentStyle
+              data-element="popover-container-content"
+              role="dialog"
+              animationState={state}
+              aria-labelledby={popoverContainerId}
+              aria-label={containerAriaLabel}
+              aria-describedby={ariaDescribedBy}
+              p="16px 24px"
+              ref={popoverContentNodeRef}
+              tabIndex={shouldCoverButton ? -1 : undefined}
+              disableAnimation={disableAnimation || reduceMotion}
+              {...filterStyledSystemPaddingProps(rest)}
+            >
+              <PopoverContainerHeaderStyle>
+                <PopoverContainerTitleStyle
+                  id={popoverContainerId}
+                  data-element="popover-container-title"
+                >
+                  {title}
+                </PopoverContainerTitleStyle>
+                {renderCloseComponent(renderCloseComponentProps)}
+              </PopoverContainerHeaderStyle>
+              {children}
+            </PopoverContainerContentStyle>
+          </Popover>
+        )
+      }
+    </Transition>
+  );
 
   return (
     <PopoverContainerWrapperStyle
@@ -253,47 +323,15 @@ export const PopoverContainer = ({
       <div ref={popoverReference}>
         {renderOpenComponent(renderOpenComponentProps)}
       </div>
-      <Transition
-        in={isOpen}
-        timeout={{ exit: 300 }}
-        appear
-        mountOnEnter
-        unmountOnExit
-        nodeRef={popoverContentNodeRef}
-      >
-        {(state: TransitionStatus) =>
-          isOpen && (
-            <Popover
-              reference={popoverReference}
-              placement={position === "right" ? "bottom-start" : "bottom-end"}
-              {...(shouldCoverButton && { middleware: popoverMiddleware })}
-            >
-              <PopoverContainerContentStyle
-                data-element="popover-container-content"
-                role="dialog"
-                animationState={state}
-                aria-labelledby={popoverContainerId}
-                aria-label={containerAriaLabel}
-                aria-describedby={ariaDescribedBy}
-                p="16px 24px"
-                ref={popoverContentNodeRef}
-                {...filterStyledSystemPaddingProps(rest)}
-              >
-                <PopoverContainerHeaderStyle>
-                  <PopoverContainerTitleStyle
-                    id={popoverContainerId}
-                    data-element="popover-container-title"
-                  >
-                    {title}
-                  </PopoverContainerTitleStyle>
-                  {renderCloseComponent(renderCloseComponentProps)}
-                </PopoverContainerHeaderStyle>
-                {children}
-              </PopoverContainerContentStyle>
-            </Popover>
-          )
-        }
-      </Transition>
+      {shouldCoverButton ? (
+        <ModalContext.Provider value={{ isAnimationComplete }}>
+          <FocusTrap wrapperRef={popoverContentNodeRef} isOpen={isOpen}>
+            {popover}
+          </FocusTrap>
+        </ModalContext.Provider>
+      ) : (
+        popover
+      )}
     </PopoverContainerWrapperStyle>
   );
 };